Full health bars, damage numbers for every hit, and bots that step off a closing wall
Three faults reported from play, each with a different shape.

Health was a uint8 on the wire while the admin allows player.maxHealth up to
1000. Raising it gave players 232 health -- 1000 truncated to a byte -- and a
red sliver of a health bar, with the server authoritative, correct, and wrong by
the time it arrived. It is a uint16 now, and tests/wire.test.ts pushes every
configurable maximum through a real Encoder/Decoder round trip so a field too
narrow for its own configured maximum fails the suite instead of reaching a
player as a wrong number.

Damage numbers only ever appeared for hits the local player landed or took,
because the server told only the attacker and the victim that a hit had
happened. Two other players trading fire, or a whole match watched after your
own elimination, showed flashes and a quietly shortening health bar. Damage is
broadcast to the room now -- no new information, since every position and health
is already in the state each client receives -- and the client draws the three
cases in different voices: amber for a hit you landed, red for one you took,
paler and smaller for somebody else's exchange.

Bots could not see the closing walls at all. Nothing about the shrink reached
the movement code, and no simulation caught it because shrinking starts at two
minutes while a bot match ends in under one. A bot chased a memory the walls had
already swallowed and stood in the edge being crushed. Goals are now clamped
into the playable strip before a route is planned, and stepping in from a wall
runs ahead of the goal, the hazard flinch and the fall steering, because a wall
does not miss and is coming regardless. Over six closing matches: time spent
flat against an edge 37.4% -> 9.4%, crush damage 11615 -> 7302.

+++ server/src/rooms/schema/PlayerState.ts
@@ -46,7 +46,14 @@ export class PlayerState extends Schema implements SyncedPlayer {
   /** -1 = facing left, 1 = facing right. */
   @type("int8") facing = 1;
 
-  @type("uint8") health: number = getPlayerConfig().maxHealth;
+  /*
+   * Two bytes, not one. Health is configurable up to 1000 and the wire held
+   * 255: an administrator who raised it got players spawning on 232 health --
+   * 1000 truncated to a byte -- with a red sliver of a health bar and no
+   * indication anything had gone wrong. The extra byte per player per patch is
+   * nothing; a setting the game silently mangles is not.
+   */
+  @type("uint16") health: number = getPlayerConfig().maxHealth;
   @type("boolean") alive = false;
   @type("boolean") onGround = false;
